Either move tests to the axon-spring module, or remove the dependency on spring for the tests still failing.
