+++ integrationtests/src/test/resources/META-INF/persistence.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  ~ Copyright (c) 2010-2020. Axon Framework
+  ~ Copyright (c) 2010-2022. Axon Framework
   ~
   ~ Licensed under the Apache License, Version 2.0 (the "License");
   ~ you may not use this file except in compliance with the License.
@@ -31,5 +31,23 @@
         <class>org.axonframework.integrationtests.polymorphic.Child1Aggregate</class>
         <class>org.axonframework.integrationtests.polymorphic.Child2Aggregate</class>
         <class>org.axonframework.integrationtests.polymorphic.SimpleAggregate</class>
+    </persistence-unit>
+    <persistence-unit name="eventStore" transaction-type="RESOURCE_LOCAL">
+        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
+        <class>org.axonframework.eventsourcing.eventstore.jpa.DomainEventEntry</class>
+        <class>org.axonframework.eventsourcing.eventstore.jpa.SnapshotEventEntry</class>
+        <class>org.axonframework.modelling.saga.repository.jpa.AssociationValueEntry</class>
+        <class>org.axonframework.modelling.saga.repository.jpa.SagaEntry</class>
+        <properties>
+            <property name="hibernate.dialect"
+                      value="org.hibernate.dialect.HSQLDialect"/>
+            <property name="hibernate.connection.driver_class"
+                      value="org.hsqldb.jdbcDriver"/>
+            <property name="hibernate.hbm2ddl.auto" value="create-drop"/>
+            <property name="hibernate.connection.username" value="sa"/>
+            <property name="hibernate.connection.password" value=""/>
+            <property name="hibernate.connection.url"
+                      value="jdbc:hsqldb:mem:addressbook"/>
+        </properties>
     </persistence-unit>
 </persistence>
