+++ messaging-jakarta/pom.xml
@@ -331,16 +331,6 @@
                         </configuration>
                     </execution>
                 </executions>
-            </plugin>            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <configuration>
-                    <trimStackTrace>false</trimStackTrace>
-                    <useFile>false</useFile>
-                    <excludes>
-                        <exclude>**/JpaTokenStoreTest.java</exclude>
-                    </excludes>
-                </configuration>
             </plugin>
         </plugins>
     </build>
