+++ messaging/src/test/resources/META-INF/persistence.xml
@@ -15,10 +15,9 @@
   -->
 
 <persistence xmlns="http://java.sun.com/xml/ns/persistence" version="1.0">
-    <persistence-unit name="eventStore" transaction-type="RESOURCE_LOCAL">
+    <persistence-unit name="tokenstore" transaction-type="RESOURCE_LOCAL">
         <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-        <class>org.axonframework.modelling.saga.repository.jpa.AssociationValueEntry</class>
-        <class>org.axonframework.modelling.saga.repository.jpa.SagaEntry</class>
+        <class>org.axonframework.eventhandling.tokenstore.jpa.TokenEntry</class>
         <properties>
             <property name="hibernate.dialect"
                       value="org.hibernate.dialect.HSQLDialect"/>
